spell 'adapter' consistently
