+++ src/main/java/com/google/maps/internal/PriceLevelAdapter.java
@@ -19,52 +19,48 @@ import com.google.gson.TypeAdapter;
 import com.google.gson.stream.JsonReader;
 import com.google.gson.stream.JsonToken;
 import com.google.gson.stream.JsonWriter;
-import com.google.maps.model.OpeningHours.Period.OpenClose.DayOfWeek;
+import com.google.maps.model.PriceLevel;
 import java.io.IOException;
 
 /**
- * This class handles conversion from JSON to {@link DayOfWeek}.
+ * This class handles conversion from JSON to {@link PriceLevel}.
  *
  * <p>Please see <a
  * href="https://google-gson.googlecode.com/svn/trunk/gson/docs/javadocs/com/google/gson/TypeAdapter.html">GSON
  * Type Adapter</a> for more detail.
  */
-public class DayOfWeekAdaptor extends TypeAdapter<DayOfWeek> {
+public class PriceLevelAdapter extends TypeAdapter<PriceLevel> {
 
   @Override
-  public DayOfWeek read(JsonReader reader) throws IOException {
+  public PriceLevel read(JsonReader reader) throws IOException {
     if (reader.peek() == JsonToken.NULL) {
       reader.nextNull();
       return null;
     }
 
     if (reader.peek() == JsonToken.NUMBER) {
-      int day = reader.nextInt();
+      int priceLevel = reader.nextInt();
 
-      switch (day) {
+      switch (priceLevel) {
         case 0:
-          return DayOfWeek.SUNDAY;
+          return PriceLevel.FREE;
         case 1:
-          return DayOfWeek.MONDAY;
+          return PriceLevel.INEXPENSIVE;
         case 2:
-          return DayOfWeek.TUESDAY;
+          return PriceLevel.MODERATE;
         case 3:
-          return DayOfWeek.WEDNESDAY;
+          return PriceLevel.EXPENSIVE;
         case 4:
-          return DayOfWeek.THURSDAY;
-        case 5:
-          return DayOfWeek.FRIDAY;
-        case 6:
-          return DayOfWeek.SATURDAY;
+          return PriceLevel.VERY_EXPENSIVE;
       }
     }
 
-    return DayOfWeek.UNKNOWN;
+    return PriceLevel.UNKNOWN;
   }
 
   /** This method is not implemented. */
   @Override
-  public void write(JsonWriter writer, DayOfWeek value) throws IOException {
+  public void write(JsonWriter writer, PriceLevel value) throws IOException {
     throw new UnsupportedOperationException("Unimplemented method");
   }
 }
